requirements.txt
pyTelegramBotAPI==4.12.0
pymongo==4.6.1
python-dotenv==1.0.0
